Refactored cursor support
