+++ external-crates/move/crates/move-analyzer/src/completions/name_chain.rs
@@ -18,10 +18,11 @@ use crate::{
         import_insertion_info, mod_defs,
     },
     symbols::{
+        cursor::{ChainCompletionKind, ChainInfo, CursorContext},
         expansion_mod_ident_to_map_key,
         types::{
-            AutoImportInsertionInfo, ChainCompletionKind, ChainInfo, CursorContext, DefInfo,
-            FunType, MemberDef, MemberDefInfo, ModuleDefs, Symbols, VariantInfo,
+            AutoImportInsertionInfo, DefInfo, FunType, MemberDef, MemberDefInfo, ModuleDefs,
+            Symbols, VariantInfo,
         },
     },
 };
